スタイル定義のファイルを作成し、それぞれ分割するよう修正
