+++ src/AppCommon.css.ts
@@ -1,0 +1,11 @@
+import makeStyles from "@material-ui/core/styles/makeStyles";
+
+const useStyles = makeStyles((theme) => ({
+    content: {
+        flexGrow: 1,
+        height: '100vh',
+        overflow: 'auto',
+    },
+}));
+
+export default useStyles;
